+++ src/com/pezz/chess/base/Network.java
@@ -1,4 +1,3 @@
-
 /*
  * Copyright (c) 2025 Gabriele Pezzini
  * License: Creative Commons Attribution-NonCommercial 4.0 International (CC BY-NC 4.0)
@@ -10,6 +9,7 @@ package com.pezz.chess.base;
 
 import java.io.File;
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.sql.Date;
 import java.util.ArrayList;
 
@@ -415,7 +415,7 @@ public class Network
       return iGameController.getPositionNote();
    }
 
-   public void deleteNoteByPositionUID(BigDecimal aPositionUID)
+   public void deleteNoteByPositionUID(BigInteger aPositionUID)
    {
       iGameController.deleteNoteByPositionUID(aPositionUID);
    }
